# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Gin + Vue + Element UI的前后端分离权限管理系统的接口文档\n添加qq群: 74520518 进入技术交流群 请备注，谢谢！",
        "title": "kubetabs API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/wenjianzhang/kubetabs/blob/master/LICENSE.md"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/v1/loginlog": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "登录日志"
                ],
                "summary": "修改登录日志",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.LoginLog"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "登录日志"
                ],
                "summary": "添加登录日志",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.LoginLog"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/loginlog/{infoId}": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "登录日志"
                ],
                "summary": "通过编码获取登录日志",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "infoId",
                        "name": "infoId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除数据",
                "tags": [
                    "登录日志"
                ],
                "summary": "批量删除登录日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "以逗号（,）分割的infoId",
                        "name": "infoId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"删除失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/loginloglist": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "登录日志"
                ],
                "summary": "登录日志列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "dictCode",
                        "name": "dictCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "dictType",
                        "name": "dictType",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页条数",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageIndex",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/operlog": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "操作日志"
                ],
                "summary": "添加操作日志",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.SysOperLog"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/operlog/{infoId}": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "登录日志"
                ],
                "summary": "通过编码获取登录日志",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "infoId",
                        "name": "infoId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/operlog/{operId}": {
            "delete": {
                "description": "删除数据",
                "tags": [
                    "操作日志"
                ],
                "summary": "批量删除操作日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "以逗号（,）分割的operId",
                        "name": "operId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"删除失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/operloglist": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "登录日志"
                ],
                "summary": "登录日志列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "dictCode",
                        "name": "dictCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "dictType",
                        "name": "dictType",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页条数",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageIndex",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.LoginLog": {
            "type": "object",
            "properties": {
                "browser": {
                    "description": "浏览器",
                    "type": "string"
                },
                "createBy": {
                    "description": "创建人",
                    "type": "string"
                },
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "dataScope": {
                    "description": "数据",
                    "type": "string"
                },
                "infoId": {
                    "description": "主键",
                    "type": "integer"
                },
                "ipaddr": {
                    "description": "ip地址",
                    "type": "string"
                },
                "isDel": {
                    "description": "是否删除",
                    "type": "string"
                },
                "loginLocation": {
                    "description": "归属地",
                    "type": "string"
                },
                "loginTime": {
                    "description": "登录时间",
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "os": {
                    "description": "系统",
                    "type": "string"
                },
                "params": {
                    "description": "参数",
                    "type": "string"
                },
                "platform": {
                    "description": "固件",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "string"
                },
                "updateBy": {
                    "description": "更新者",
                    "type": "string"
                },
                "updateTime": {
                    "description": "更新时间",
                    "type": "string"
                },
                "userName": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据集",
                    "type": "object"
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string"
                }
            }
        },
        "models.SysOperLog": {
            "type": "object",
            "properties": {
                "businessType": {
                    "description": "操作类型",
                    "type": "string"
                },
                "businessTypes": {
                    "type": "string"
                },
                "createBy": {
                    "description": "创建人",
                    "type": "string"
                },
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "dataScope": {
                    "description": "数据",
                    "type": "string"
                },
                "deptName": {
                    "description": "部门名称",
                    "type": "string"
                },
                "isDel": {
                    "description": "是否删除",
                    "type": "string"
                },
                "jsonResult": {
                    "description": "返回数据",
                    "type": "string"
                },
                "latencyime": {
                    "description": "耗时",
                    "type": "string"
                },
                "method": {
                    "description": "函数",
                    "type": "string"
                },
                "operId": {
                    "description": "日志编码",
                    "type": "integer"
                },
                "operIp": {
                    "description": "客户端ip",
                    "type": "string"
                },
                "operLocation": {
                    "description": "访问位置",
                    "type": "string"
                },
                "operName": {
                    "description": "操作者",
                    "type": "string"
                },
                "operParam": {
                    "description": "请求参数",
                    "type": "string"
                },
                "operTime": {
                    "description": "操作时间",
                    "type": "string"
                },
                "operUrl": {
                    "description": "访问地址",
                    "type": "string"
                },
                "operatorType": {
                    "description": "操作类型",
                    "type": "string"
                },
                "params": {
                    "description": "参数",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "requestMethod": {
                    "description": "请求方式",
                    "type": "string"
                },
                "status": {
                    "description": "操作状态",
                    "type": "string"
                },
                "title": {
                    "description": "操作模块",
                    "type": "string"
                },
                "updateBy": {
                    "description": "更新者",
                    "type": "string"
                },
                "updateTime": {
                    "description": "更新时间",
                    "type": "string"
                },
                "userAgent": {
                    "description": "us",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}